+++ ru.taximaxim.codekeeper.core.tests/src/main/resources/ru/taximaxim/codekeeper/core/drop_ms_column_options_original.sql
@@ -7,5 +7,7 @@ CREATE TABLE [dbo].[table1](
     [c2] [int] NOT NULL,
     [c3] [uniqueidentifier] NULL ROWGUIDCOL,
     [c4] AS ([c1]+[c2]) PERSISTED,
-    [c5] [varchar] (50) COLLATE Cyrillic_General_CI_AS SPARSE NULL)
+    [c5] [varchar] (50) COLLATE Cyrillic_General_CI_AS SPARSE NULL,
+    [c6] [varchar] (100) MASKED WITH (FUNCTION = 'partial(1, "XXX''XXXX", 0)') NULL,
+    [c7] [int] IDENTITY(1,1) NOT FOR REPLICATION NOT NULL)
 GO